+++ src/graph/executor/algo/ShortestPathBase.h
@@ -4,7 +4,6 @@
 #ifndef GRAPH_EXECUTOR_QUERY_SHORTESTPATHBASE_H_
 #define GRAPH_EXECUTOR_QUERY_SHORTESTPATHBASE_H_
 
-#include "graph/executor/StorageAccessExecutor.h"
 #include "graph/planner/plan/Algo.h"
 
 using nebula::storage::StorageRpcResponse;
@@ -15,11 +14,19 @@ namespace nebula {
 namespace graph {
 class ShortestPathBase {
  public:
-  ShortestPathBase(const PlanNode* node, QueryContext* qctx) {
-    pathNode_ = asNode<ShortestPath>(node);
+  ShortestPathBase(const ShortestPath* node,
+                   QueryContext* qctx,
+                   std::unordered_map<std::string, std::string>* stats)
+      : pathNode_(node), qctx_(qctx), stats_(stats) {
+    singleShortest_ = node->singleShortest();
+    maxStep_ = node->stepRange()->max();
   }
 
-  folly::Future<Status> execute();
+  virtual ~ShortestPathBase() {}
+
+  virtual folly::Future<Status> execute(const std::unordered_set<Value>& startVids,
+                                        const std::unordered_set<Value>& endVids,
+                                        DataSet* result) = 0;
 
   using DstVid = Value;
   // start vid of the path
@@ -30,16 +37,8 @@ class ShortestPathBase {
   using CustomStep = Row;
   using HalfPath = std::vector<std::unordered_map<DstVid, std::vector<CustomStep>>>;
 
- private:
-  folly::Future<Status> getNeighbors(size_t rowNum, bool reverse);
-
-  folly::Future<Status> shortestPath(size_t rowNum, size_t stepNum);
-
-  folly::Future<Status> handleResponse(size_t rowNum, size_t stepNum);
-
+ protected:
   Status handlePropResp(PropRpcResponse&& resps, std::vector<Value>& vertices);
-
-  Status buildPath(size_t rowNum, RpcResponse&& resp, bool reverse);
 
   folly::Future<Status> getMeetVidsProps(const std::vector<Value>& meetVids,
                                          std::vector<Value>& meetVertices);
@@ -48,8 +47,53 @@ class ShortestPathBase {
 
   std::vector<Row> createLeftPath(size_t rowNum, const Value& meetVid);
 
- private:
+  Status handleErrorCode(nebula::cpp2::ErrorCode code, PartitionID partId) const;
+
+  template <typename RESP>
+  void addStats(RESP& resp, std::unordered_map<std::string, std::string>* stats) const {
+    auto& hostLatency = resp.hostLatency();
+    for (size_t i = 0; i < hostLatency.size(); ++i) {
+      auto& info = hostLatency[i];
+      stats->emplace(folly::sformat("{} exec/total", std::get<0>(info).toString()),
+                     folly::sformat("{}(us)/{}(us)", std::get<1>(info), std::get<2>(info)));
+      auto detail = getStorageDetail(resp.responses()[i].result_ref()->latency_detail_us_ref());
+      if (!detail.empty()) {
+        stats->emplace("storage_detail", detail);
+      }
+    }
+  }
+
+  template <typename Resp>
+  StatusOr<Result::State> handleCompleteness(const storage::StorageRpcResponse<Resp>& rpcResp,
+                                             bool isPartialSuccessAccepted) const {
+    auto completeness = rpcResp.completeness();
+    if (completeness != 100) {
+      const auto& failedCodes = rpcResp.failedParts();
+      for (auto it = failedCodes.begin(); it != failedCodes.end(); it++) {
+        LOG(ERROR) << " failed, error " << apache::thrift::util::enumNameSafe(it->second)
+                   << ", part " << it->first;
+      }
+      // cannot execute at all, or partial success is not accepted
+      if (completeness == 0 || !isPartialSuccessAccepted) {
+        if (failedCodes.size() > 0) {
+          return handleErrorCode(failedCodes.begin()->second, failedCodes.begin()->first);
+        }
+        return Status::Error("Request to storage failed, without failedCodes.");
+      }
+      // partial success is accepted
+      qctx_->setPartialSuccess();
+      return Result::State::kPartialSuccess;
+    }
+    return Result::State::kSuccess;
+  }
+
+  std::string getStorageDetail(
+      apache::thrift::optional_field_ref<const std::map<std::string, int32_t>&> ref) const;
+
+ protected:
   const ShortestPath* pathNode_{nullptr};
+  QueryContext* qctx_{nullptr};
+  std::unordered_map<std::string, std::string>* stats_{nullptr};
   size_t maxStep_;
   bool singleShortest_{true};
