+++ tests/tck/features/match/AllShortestPaths.feature
@@ -1,6 +1,7 @@
 # Copyright (c) 2022 vesoft inc. All rights reserved.
 #
 # This source code is licensed under Apache 2.0 License.
+@jmq
 Feature: allShortestPaths
 
   Background:
